+++ docs/griptape-framework/drivers/src/event_listener_drivers_1.py
@@ -1,25 +1,24 @@
 import os
 
 from griptape.drivers import AmazonSqsEventListenerDriver
-from griptape.events import (
-    EventListener,
-)
+from griptape.events import EventListener, event_bus
 from griptape.rules import Rule
 from griptape.structures import Agent
 
-agent = Agent(
-    rules=[
-        Rule(value="You will be provided with a block of text, and your task is to extract a list of keywords from it.")
-    ],
-    event_listeners=[
+event_bus.add_event_listeners(
+    [
         EventListener(
-            handler=lambda event: {  # You can optionally use the handler to transform the event payload before sending it to the Driver
-                "event": event.to_dict(),
-            },
             driver=AmazonSqsEventListenerDriver(
                 queue_url=os.environ["AMAZON_SQS_QUEUE_URL"],
             ),
         ),
+    ]
+)
+
+
+agent = Agent(
+    rules=[
+        Rule(value="You will be provided with a block of text, and your task is to extract a list of keywords from it.")
     ],
 )