+++ docs/griptape-framework/misc/src/events_1.py
@@ -7,6 +7,7 @@ from griptape.events import (
     StartActionsSubtaskEvent,
     StartPromptEvent,
     StartTaskEvent,
+    event_bus,
 )
 from griptape.structures import Agent
 
@@ -15,8 +16,8 @@ def handler(event: BaseEvent) -> None:
     print(event.__class__)
 
 
-agent = Agent(
-    event_listeners=[
+event_bus.add_event_listeners(
+    [
         EventListener(
             handler,
             event_types=[
@@ -30,5 +31,7 @@ agent = Agent(
         )
     ]
 )
+
+agent = Agent()
 
 agent.run("tell me about griptape")