+++ docs/griptape-framework/misc/src/events_3.py
@@ -2,19 +2,22 @@ from typing import cast
 
 from griptape.config import OpenAiStructureConfig
 from griptape.drivers import OpenAiChatPromptDriver
-from griptape.events import CompletionChunkEvent, EventListener
+from griptape.events import CompletionChunkEvent, EventListener, event_bus
 from griptape.structures import Pipeline
 from griptape.tasks import ToolkitTask
 from griptape.tools import TaskMemoryClient, WebScraper
 
-pipeline = Pipeline(
-    config=OpenAiStructureConfig(prompt_driver=OpenAiChatPromptDriver(model="gpt-4o", stream=True)),
-    event_listeners=[
+event_bus.add_event_listeners(
+    [
         EventListener(
             lambda e: print(cast(CompletionChunkEvent, e).token, end="", flush=True),
             event_types=[CompletionChunkEvent],
         )
-    ],
+    ]
+)
+
+pipeline = Pipeline(
+    config=OpenAiStructureConfig(prompt_driver=OpenAiChatPromptDriver(model="gpt-4o", stream=True)),
 )
 
 pipeline.add_tasks(