+++ griptape/engines/rag/rag_context.py
@@ -19,10 +19,10 @@ class RagContext(SerializableMixin):
     Attributes:
         query: Query provided by the user.
         module_configs: Dictionary of module configs. First key should be a module name and the second a dictionary of config parameters.
-        before_query: An optional list of strings to add before the query in generation modules.
-        after_query: An optional list of strings to add after the query in generation modules.
-        text_chunks: A list of text chunks to pass around from the retrieval stage to the generation stage.
-        output: Final output from the generation stage.
+        before_query: An optional list of strings to add before the query in response modules.
+        after_query: An optional list of strings to add after the query in response modules.
+        text_chunks: A list of text chunks to pass around from the retrieval stage to the response stage.
+        output: Final output from the response stage.
     """
 
     query: str = field(metadata={"serializable": True})