+++ CHANGELOG.md
@@ -11,8 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to set custom schema properties on Tool Activities via `extra_schema_properties`.
 - Parameter `structure` to `BaseTask`.
 - Method `try_find_task` to `Structure`.
+- `TranslateQueryRagModule` `RagEngine` module for translating input queries.
+- Global event bus, `griptape.events.event_bus`, for publishing and subscribing to events.
 
 ### Changed
+- **BREAKING**: Removed all uses of `EventPublisherMixin` in favor of `event_bus`.
+- **BREAKING**: Removed `EventPublisherMixin`.
 - `BaseTask.add_parent/child` will now call `self.structure.add_task` if possible.
 
 ## [0.29.0] - 2024-07-30